requirements.txt
# Файл списка зависимостей
# Команда для установки пакетов из этого файла:
# pip install -r requirements.txt
# Команда для установки с обновлением:
# pip install --upgrade -r requirements.txt
# Альтернативная команда установки пакетов:
# pip install playwright pytest pytest-html allure-pytest requests locust
# Команда формирования этого документа (заменяет старый т.е. этот):
# pip freeze > requirements.txt

# базовые библиотеки тестирования
playwright==1.51.0 # основной фреймворк тестирования ui
pytest==8.3.5 # библиотека управления тестами

# отчёты
pytest-html==4.1.1 # отчеты пайтест в формате хтмл
allure-pytest==2.14.0 # отчеты аллюр и пайтест

# апи
requests==2.32.3 # хттп запросы

# нагрузочное
locust==2.35.0 # простое нагрузочное тестирование
